// frontend/src/app/api/model/transactie.ts
/**
 * BankSim API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Partij } from './partij';
import { TransactieType } from './transactie-type';


export interface Transactie { 
    id: string;
    datum: string;
    tijdstip: string;
    uitgevoerdOp: string;
    tegenNaam: string;
    tegenIban?: string;
    /**
     * Bedrag in euro\'s met maximaal 2 decimalen; punt als decimaalteken
     */
    bedrag: string;
    type: TransactieType;
    typeLabel: string;
    rekeninghouder: string;
    van: Partij;
    naar: Partij;
    omschrijving?: string;
    betalingskenmerk?: string;
    extraOmschrijving?: string;
}